pyproject.toml
[tool.poetry]
name = "tireshop"
version = "0.1.0"
description = ""
authors = ["Augusto Bendlin <augusto.bendlin1997@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
